Refactor the code to split out the storage db counters
